move DayWeather class to separate file
